Refactor: extract the TFS pull request number extraction to a class so later we can test it
