+++ src/PullRequestReleaseNotes/Providers/TfsPullRequestProvider.cs
@@ -187,11 +187,7 @@ namespace PullRequestReleaseNotes.Providers
 
         private static int? ExtractPullRequestNumber(string commitMessage)
         {
-            var pattern = new Regex(@"[Merge pull request|Merged PR] (?<pullRequestNumber>\d+).*");
-            var match = pattern.Match(commitMessage);
-            if (match.Groups.Count <= 0 || !match.Groups["pullRequestNumber"].Success)
-                return null;
-            return int.Parse(match.Groups["pullRequestNumber"].Value);
+            return TfsPullRequestNumberExtractor.Extract(commitMessage);
         }
     }
 }
